Add GitHub::User superclass to GitHub::Follower and GitHub::Following to DRY
