+++ app/models/git_hub/follower.rb
@@ -1,17 +1,10 @@
 module GitHub
-  class Follower
-    attr_reader :login, :url
-
-    def initialize(attributes)
-      @login = attributes[:login]
-      @url = attributes[:html_url]
-    end
-
+  class Follower < GitHub::User
     def self.list_all(gh_token)
       followers = GitHubService.new(gh_token).followers
 
       followers.map do |follower_hash|
-        Follower.new(follower_hash)
+        GitHub::User.new(follower_hash)
       end
     end
   end
