+++ app/models/git_hub/following.rb
@@ -1,17 +1,12 @@
 module GitHub
-  class Following
-    attr_reader :login, :url
-
-    def initialize(attributes)
-      @login = attributes[:login]
-      @url = attributes[:html_url]
-    end
-
+  # User class to combine follwoer/following
+  # or GH user superclass
+  class Following < GitHub::User
     def self.list_all(gh_token)
       followings = GitHubService.new(gh_token).following
 
       followings.map do |following_hash|
-        Following.new(following_hash)
+        GitHub::User.new(following_hash)
       end
     end
   end
