+++ app/models/git_hub/user.rb
@@ -1,0 +1,10 @@
+module GitHub
+  class User
+    attr_reader :login, :url
+
+    def initialize(attributes)
+      @login = attributes[:login]
+      @url = attributes[:html_url]
+    end
+  end
+end
